requirements.txt
anthropic>=0.39.0
praw>=7.7.0
yfinance>=0.2.40
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.31.0
